systems/rss-digest-monitor/requirements.txt
# RSS Digest Monitor Dependencies

# RSS/Atom feed parsing - universal parser for all feed formats
feedparser==6.0.11

# Date parsing - handles various datetime formats robustly
python-dateutil==2.8.2

# Optional: HTTP client with better timeout/retry than requests
# (feedparser uses requests internally, but this is here for reference)
# httpx==0.27.0
